+++ RaytraceGroundUp/Chapter29.4_linux/Whitted.h
@@ -16,8 +16,8 @@ class Whitted: public Tracer {
         virtual RGBColor    
         trace_ray(const Ray ray, const int depth) const;
         
-		virtual RGBColor
-		trace_ray(const Ray ray, double& tmin, const int depth) const;
+        virtual RGBColor
+        trace_ray(const Ray ray, float& tmin, const int depth) const;
 };
 
 #endif
