+++ RaytraceGroundUp/Chapter31.1.2_linux_module/Tracers/GlobalTrace.cpp
@@ -1,7 +1,7 @@
-// 	Copyright (C) Kevin Suffern 2000-2007.
-//	This C++ code is for non-commercial purposes only.
-//	This C++ code is licensed under the GNU General Public License Version 2.
-//	See the file COPYING.txt for the full license.
+//     Copyright (C) Kevin Suffern 2000-2007.
+//    This C++ code is for non-commercial purposes only.
+//    This C++ code is licensed under the GNU General Public License Version 2.
+//    See the file COPYING.txt for the full license.
 
 
 #include "GlobalTrace.h"
@@ -12,14 +12,14 @@
 // -------------------------------------------------------------------- default constructor
 
 GlobalTrace::GlobalTrace(void)
-	: Tracer()
+    : Tracer()
 {}
 
 
 // -------------------------------------------------------------------- default constructor
 
 GlobalTrace::GlobalTrace(World* _worldPtr)
-	: Tracer(_worldPtr)
+    : Tracer(_worldPtr)
 {}
 
 
@@ -27,22 +27,22 @@ GlobalTrace::GlobalTrace(World* _worldPtr)
 
 RGBColor
 GlobalTrace::trace_ray(const Ray ray, const int depth) const {
-	if (depth > world_ptr->vp.max_depth)
-		return (black);
-	else {
-		ShadeRec sr(world_ptr->hit_objects(ray));
+    if (depth > world_ptr->vp.max_depth)
+        return (black);
+    else {
+        ShadeRec sr(world_ptr->hit_objects(ray));
 
-		if (sr.hit_an_object) {
-			sr.depth = depth;
-			sr.ray = ray;
+        if (sr.hit_an_object) {
+            sr.depth = depth;
+            sr.ray = ray;
 
             // 该函数与PathTrace::trace_ray() 函数的唯一差别在于：
             // 该函数调用了material_ptr->global_shade()函数而非material_ptr->shade()函数。
-			return (sr.material_ptr->global_shade(sr));
-		}
-		else
-			return (world_ptr->background_color);
-	}
+            return (sr.material_ptr->global_shade(sr));
+        }
+        else
+            return (world_ptr->background_color);
+    }
 }
 
 
@@ -51,23 +51,23 @@ GlobalTrace::trace_ray(const Ray ray, const int depth) const {
 // for color filtering
 
 RGBColor
-GlobalTrace::trace_ray(const Ray ray, double& tmin, const int depth) const {
-	if (depth > world_ptr->vp.max_depth)
-		return (black);
-	else {
-		ShadeRec sr(world_ptr->hit_objects(ray));
+GlobalTrace::trace_ray(const Ray ray, float& tmin, const int depth) const {
+    if (depth > world_ptr->vp.max_depth)
+        return (black);
+    else {
+        ShadeRec sr(world_ptr->hit_objects(ray));
 
-		if (sr.hit_an_object) {
-			sr.depth 	= depth;
-			sr.ray 		= ray;
-			tmin		= sr.t;     // required for colored transparency
+        if (sr.hit_an_object) {
+            sr.depth     = depth;
+            sr.ray         = ray;
+            tmin        = sr.t;     // required for colored transparency
             // 该函数与PathTrace::trace_ray() 函数的唯一差别在于：
             // 该函数调用了material_ptr->global_shade()函数而非material_ptr->shade()函数。
-			return (sr.material_ptr->global_shade(sr));
-		}
-		else{
-			tmin = kHugeValue;
-			return (world_ptr->background_color);
-		}
-	}
+            return (sr.material_ptr->global_shade(sr));
+        }
+        else{
+            tmin = kHugeValue;
+            return (world_ptr->background_color);
+        }
+    }
 }
